Worked around issue getting consul_ui working
Worked around an issue with the current Consul cookbook.

The Consul cookbook uses ark to install the UI component, however ark's
documentation isn't clear that both the `path` and `home_dir` attribute
are NOT consumer configurable.  Instead these properties are derived
from other values (namely the `prefix_root` and `prefix_home`
attributes).

We'll run with this until a proper fix is applied to the Consul
cookbook.